embeddings/requirements.txt
langchain
python-dotenv
openai
tiktoken